@@ -68,10 +69,10 @@
     "jest-transform-stub": "^2.0.0",
     "lerna": "^5.3.0",
     "process": "0.11.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.3.0",
-    "stream-browserify": "3.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
+    "stream-browserify": "^3.0.0",
     "style-loader": "^3.3.0",
     "styled-components": "^5.3.1",
     "terser-webpack-plugin": "5.1.1",
@@ -79,41 +80,51 @@
     "typescript": "^4.3.5",
     "url-loader": "^4.1.1",
     "util": "0.12.3",
-    "webpack": "5.24.4",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^3.11.2"
+    "webpack": "5.75.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^3.11.3"
   },
   "dependencies": {
-    "@connect2ic/core": "^0.2.0-beta.5",
-    "@connect2ic/react": "^0.2.0-beta.5",
+    "@connect2ic/core": "^0.2.0-beta.21",
+    "@connect2ic/react": "^0.2.0-beta.21",
     "@date-io/moment": "^2.14.0",
     "@dfinity/auth-client": "^0.10.4",
     "@dfinity/authentication": "^0.10.4",
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
+    "@emotion/core": "^11.0.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@hookform/resolvers": "^2.9.3",
-    "@material-ui/lab": "^4.0.0-alpha.61",
     "@mui/icons-material": "^5.8.4",
-    "@mui/lab": "^5.0.0-alpha.94",
-    "@mui/material": "^5.8.5",
-    "@mui/x-date-pickers": "^5.0.0-alpha.7",
+    "@mui/lab": "^5.0.0-alpha.109",
+    "@mui/material": "^5.10.15",
+    "@mui/x-date-pickers": "^5.0.9",
+    "@origyn-sa/mintjs": "^1.0.0-alpha.4",
+    "@origyn-sa/origyn-art-ui": "^1.2.13",
     "@svgr/webpack": "^6.2.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@types/react": "^17.0.19",
     "@types/react-dom": "^17.0.9",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
+    "bip39": "^3.0.4",
     "buffer-crc32": "^0.2.13",
+    "cross-fetch": "^3.1.5",
+    "crypto-browserify": "^3.12.0",
     "crypto-js": "^4.1.1",
     "ethereum-blockies-base64": "^1.0.2",
+    "hdkey": "^2.0.1",
+    "history": "^5.3.0",
     "ic-stoic-identity": "^2.0.0",
     "json-bigint": "^1.0.0",
     "moment": "^2.29.3",
+    "mrmime": "^1.0.1",
+    "mui-file-dropzone": "^4.0.2",
     "notistack": "^2.0.5",
     "react-hook-form": "^7.33.0",
     "react-is": "^18.2.0",
     "react-json-to-table": "^0.1.7",
     "react-toastify": "^8.1.0",
     "react-tooltip": "^4.2.21",
+    "tss-react": "^4.4.4",
     "yup": "^0.32.11"
   },
   "browserslist": [
