@@ -1,5 +1,5 @@
-import { SnackbarProvider } from 'notistack';
 import React from 'react';
+import { SnackbarProvider } from 'notistack';
 import { HashRouter, Route, Routes } from 'react-router-dom';
 import { NFTPage } from '@dapp/features-sales-escrows';
 import { SiteProvider } from '@dapp/features-theme';
@@ -7,6 +7,7 @@ import { TokensContextProvider } from '@dapp/features-tokens-provider';
 import Marketplace from './pages/Marketplace';
 import { Layout } from '@dapp/features-components';
 import { AuthProvider, SessionProvider } from '@dapp/features-authentication';
+import { MarketplaceProvider } from './components/context';
 
 const App = () => (
   <HashRouter>
@@ -15,12 +16,14 @@ const App = () => (
         <TokensContextProvider>
           <AuthProvider>
             <SnackbarProvider maxSnack={3}>
-              <Layout>
-                <Routes>
-                  <Route path="/" element={<Marketplace />} />
-                  <Route path="/:nft_id" element={<NFTPage />} />
-                </Routes>
-              </Layout>
+              <MarketplaceProvider>
+                <Layout>
+                  <Routes>
+                    <Route path="/" element={<Marketplace />} />
+                    <Route path="/:nft_id" element={<NFTPage />} />
+                  </Routes>
+                </Layout>
+              </MarketplaceProvider>
             </SnackbarProvider>
           </AuthProvider>
         </TokensContextProvider>
