@@ -6,21 +6,26 @@ import { SiteProvider } from '@dapp/features-theme';
 import 'react-toastify/dist/ReactToastify.css';
 import { TokensContextProvider } from '@dapp/features-tokens-provider';
 import { SnackbarProvider } from 'notistack';
+import { MetaProvider } from './components/nftTabs/context';
 import { AuthProvider, SessionProvider } from '@dapp/features-authentication';
+import { GlobalStyle } from '@origyn-sa/origyn-art-ui'
 
 const App = () => {
   return (
     <HashRouter>
+      <GlobalStyle />
       <SiteProvider>
         <SessionProvider>
           <TokensContextProvider>
             <AuthProvider>
               <SnackbarProvider maxSnack={3}>
-                <Layout>
-                  <Routes>
-                    <Route path="*" element={<Home />} />
-                  </Routes>
-                </Layout>
+                  <Layout>
+                    <MetaProvider>
+                    <Routes>
+                      <Route path="*" element={<Home />} />
+                    </Routes>
+                    </MetaProvider>
+                  </Layout>
               </SnackbarProvider>
             </AuthProvider>
           </TokensContextProvider>
