@@ -1,40 +1,48 @@
-import React, { useContext, useEffect, useState } from 'react';
-import { AuthContext, useRoute, useSessionContext } from '@dapp/features-authentication';
+import React, { useEffect, useState } from 'react';
+import { useRoute } from '@dapp/features-authentication';
 import NFTInfo from '../NFTInfo';
-import { isLocal } from '@dapp/utils';
+import { OrigynClient, getNftCollectionMeta } from '@origyn-sa/mintjs';
+import { Container } from '@origyn-sa/origyn-art-ui';
 
 const Home = () => {
-  const { localDevelopment } = useSessionContext();
-  const { tokenId, principal } = useContext(AuthContext);
+  // const { actor } = useContext(AuthContext);
+  const [tokenId, setTokenId] = useState();
   const [NFTData, setNFTData] = useState();
   const [canisterId, setCanisterId] = useState('');
 
+  const nftCollection = async () => {
+    const route = await useRoute();
+
+    OrigynClient.getInstance().init(true, route.canisterId);
+    const response = await getNftCollectionMeta([]);
+    const collectionNFT = response.ok;
+    const obj_token_ids: any = collectionNFT.token_ids[0];
+
+    // In case we have URL with tokenID and we change canister,
+    // We need to check if the tokenID is in the new canister
+    // If not, we need to clear the URL and show the first tokenID in the new Canister
+    // if (!obj_token_ids.includes(tokenId) && tokenId !== '') {
+    //   let Url = window.location.href;
+    //   Url = Url.replace(tokenId, obj_token_ids[0]);
+    //   window.location.href = Url;
+    // }
+    return obj_token_ids;
+  };
+
   const getData = async () => {
-    const url =
-      isLocal() && localDevelopment
-        ? `http://${canisterId}.localhost:8000`
-        : `https://${canisterId}.raw.ic0.app`;
     if (tokenId) {
       try {
-        const response = await fetch(url + '/-/${tokenId}/info');
+        const response = await fetch(`https://${canisterId}.raw.ic0.app/-/${tokenId}/info`);
         const result = await response.text();
-        if (result.search('"is_soulbound":,')) {
-          setNFTData(JSON.parse(result.replace('"is_soulbound":,', '')));
-        } else {
-          setNFTData(JSON.parse(result));
-        }
+        setNFTData(JSON.parse(result.replace(':,', ':"",')));
       } catch (err) {
         console.log(err);
       }
     } else {
       try {
-        const response = await fetch(url + '/collection/info');
+        const response = await fetch(`https://${canisterId}.raw.ic0.app/collection/info`);
         const result = await response.text();
-        if (result.search('"is_soulbound":,')) {
-          setNFTData(JSON.parse(result.replace('"is_soulbound":,', '')));
-        } else {
-          setNFTData(JSON.parse(result));
-        }
+        setNFTData(JSON.parse(result.replace(':,', ':"",')));
       } catch (err) {
         console.log(err);
       }
@@ -42,36 +50,20 @@ const Home = () => {
   };
 
   useEffect(() => {
-    useRoute().then(({ canisterId }) => {
+    document.title = 'Origyn NFT data browser';
+    useRoute().then(({ canisterId, tokenId }) => {
       setCanisterId(canisterId);
+      setTokenId(tokenId);
     });
   }, []);
+
   useEffect(() => {
     if (canisterId) {
+      nftCollection();
       getData();
     }
   }, [canisterId]);
 
-  return (
-    <div>
-      <div>
-        <p>
-          Canister ID: <b>{canisterId}</b>
-        </p>
-        <p>
-          Token(NFT) ID: <b>{tokenId}</b>
-        </p>
-        <p>
-          You principal: <b>{principal?.toText()}</b>
-        </p>
-        <div>
-          <p>
-            <b>NFT Data:</b>
-          </p>
-          {NFTData ? <NFTInfo metadata={NFTData} /> : null}
-        </div>
-      </div>
-    </div>
-  );
+  return <Container>{NFTData ? <NFTInfo metadata={NFTData} /> : null}</Container>;
 };
 export default Home;
