@@ -1,75 +1,42 @@
 import * as React from 'react';
-import Box from '@mui/material/Box';
-import Tab from '@mui/material/Tab';
-
-import Tabs from '@mui/material/Tabs';
 import TreeTab from '../../components/nftTabs/TreeTab';
 import RawTab from '../../components/nftTabs/RawTab';
 import FormTab from '../../components/nftTabs/FormTab';
-import Container from '@mui/material/Container';
-import Typography from '@mui/material/Typography';
 import JSONTab from '../../components/nftTabs/JSONTab';
+import Write from '../../components/nftTabs/Write';
+import { SecondaryNav } from '@origyn-sa/origyn-art-ui';
+import { useContext } from 'react';
+import { AuthContext } from '../../../../../features/authentication';
+import { useDialog } from '@connect2ic/react';
 
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}
-
-function TabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props;
+const NFTInfo = ({ metadata }: any) => {
+  const { principal, handleLogOut } = useContext(AuthContext);
+  const { open } = useDialog();
 
   return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box sx={{}}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
+    <>
+      <SecondaryNav
+        title="Certificates Data"
+        tabs={[
+          { title: 'Form', id: 'Form' },
+          { title: 'Tree', id: 'Tree' },
+          { title: 'Raw', id: 'Raw' },
+          { title: 'JSON', id: 'JSON' },
+          { title: 'Write', id: 'Write'}
+        ]}
+        content={[
+          <FormTab metadata={metadata} key="formTab" />,
+          <TreeTab metadata={metadata} key="treeTab" />,
+          <RawTab metadata={metadata} key="rawTab" />,
+          <JSONTab metadata={metadata} key="jsonTab" />,
+          <Write key="writeTab" />
+        ]}
+        onLogOut={handleLogOut}
+        onConnect={open}
+        principal={principal?.toText() === '2vxsx-fae' ? '' : principal?.toText()}
+      />
+    </>
   );
-}
+};
 
-const NFTInfo = ({metadata}: any) => {
-    const [value, setValue] = React.useState(1);
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue);
-  };
-    
-    return (
-      <Container maxWidth="xl">
-
-    <Box sx={{ width: '100%' }}>
-      <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
-        <Tabs value={value} onChange={handleChange} aria-label="basic tabs example">
-          <Tab label="Form"  />
-          <Tab label="Tree"  />
-          <Tab label="Raw"  />
-          <Tab label="JSON"  />
-        </Tabs>
-      </Box>
-      <TabPanel value={value} index={0}>
-      <FormTab metadata={metadata} />
-      </TabPanel>
-      <TabPanel value={value} index={1}>
-      <TreeTab metadata={metadata} />
-      </TabPanel>
-      <TabPanel value={value} index={2}>
-        <RawTab metadata={metadata} />
-      </TabPanel>
-      <TabPanel value={value} index={3}>
-        <JSONTab metadata={metadata} />
-      </TabPanel>
-    </Box>
-      </Container>
-    );
-  };
-  
-  export default NFTInfo;
+export default NFTInfo;
