@@ -1,8 +1,8 @@
 const webpack = require('../../../webpack.config');
 
 const appConfig = {
-  name: 'nftData',
-  openPage: '-/baycdev/-/bayc-1/-/nftData',
+  name: 'data',
+  openPage: '-/bmdev/-/bm-1/-/data',
 };
 module.exports = (env, argv) => {
   const res = webpack(env, argv, appConfig);
