@@ -4,7 +4,8 @@ import { Preloader } from '@dapp/features-components';
 import { Principal } from '@dfinity/principal';
 import { createContext, useContext } from 'react';
 import { AuthContextType } from '../types';
-import { useSessionContext } from './SessionProvider';
+import styled from 'styled-components';
+import PropTypes from 'prop-types';
 
 export const AuthContext = createContext<AuthContextType>({
   loggedIn: false,
@@ -14,28 +15,60 @@ export const AuthContext = createContext<AuthContextType>({
 export const useAuthContext = () => useContext(AuthContext);
 
 export const useAuth = () => {
-  const { localDevelopment, canisterId } = useSessionContext();
-
   const {
     activeProvider: activeWalletProvider,
     principal,
     disconnect: handleLogOut,
     isInitializing,
     isConnected,
+    ...other
   } = useConnect();
-
   const [actor] = useCanister('nft');
 
   return {
     activeWalletProvider,
     actor,
-    canisterId,
     handleLogOut,
+    other,
     isLoading: isInitializing,
-    localDevelopment,
     loggedIn: isConnected,
     principal: principal?.length > 0 ? Principal.fromText(principal) : Principal.anonymous(),
   };
+};
+
+const StyledPlugNotification = styled.div`
+  position: absolute;
+  top: 50px;
+  right: 50px;
+  border-width: 4px;
+  border-style: solid;
+  padding: 16px;
+  box-sizing: border-box;
+  max-width: calc(100% - 100px);
+  border-image: linear-gradient(
+      96.38deg,
+      rgb(255, 231, 1) -0.67%,
+      rgb(250, 81, 211) 31.53%,
+      rgb(16, 217, 237) 61.61%,
+      rgb(95, 255, 96) 100.67%
+    )
+    1;
+  @media (max-width: 767px) {
+    top: 80px;
+    right: 8px;
+    max-width: calc(100% - 16px);
+  }
+`;
+
+const PlugNotification = () => {
+  return (
+    <StyledPlugNotification>
+      <p>
+        If you are using Plug wallet,
+        <br /> <b>make sure you have logged-in to your Plug wallet extension.</b>
+      </p>
+    </StyledPlugNotification>
+  );
 };
 
 export const AuthProvider = ({ children }) => {
@@ -44,10 +77,17 @@ export const AuthProvider = ({ children }) => {
     <>
       <ConnectDialog />
       {auth.isLoading ? (
-        <Preloader />
+        <>
+          <Preloader width="100%" />
+          {window?.ic?.plug?.sessionManager?.initialized ? <PlugNotification /> : ''}
+        </>
       ) : (
         <AuthContext.Provider value={auth}>{children}</AuthContext.Provider>
       )}
     </>
   );
+};
+
+AuthProvider.propTypes = {
+  children: PropTypes.object
 };
