@@ -1,51 +1,83 @@
 import React, { useContext, useEffect } from 'react';
-import { AuthContext } from '@dapp/features-authentication';
-import { Box, Typography } from '@mui/material';
-import TextField from '@mui/material/TextField';
-import Autocomplete from '@mui/material/Autocomplete';
-import Paper from '@mui/material/Paper';
-import FormControl from '@mui/material/FormControl';
+import { AuthContext, useRoute } from '@dapp/features-authentication';
 import { collectionName } from '@dapp/utils';
-// Preloader
-import { CircularProgress } from '@mui/material';
+import { getNftCollectionMeta, OrigynClient } from '@origyn-sa/mintjs';
+import {
+  Container,
+  Card,
+  Select,
+  HR,
+  Grid,
+  Icons,
+  Flex,
+  Button,
+  TextInput,
+} from '@origyn-sa/origyn-art-ui';
 
-export const SearchbarNft = (props : any) => {
+export const SearchbarNft = (props: any) => {
+  const { actor } = useContext(AuthContext);
+  const [tokenId, setTokenId] = React.useState('');
+  const [canisterId, setCanisterId] = React.useState('');
+  const [selectTokenIds, setSelectTokenIds] = React.useState<any>(['']);
 
-  const { tokenId, actor } = useContext(AuthContext);
-  const [selectTokenIds, setSelectTokenIds] = React.useState(['']);
-  const [idsNumber, setIdsNumber] = React.useState('');
+  // TODO: uncomment when idsNumber is being used
+  // const [idsNumber, setIdsNumber] = React.useState('');
 
-  const handleSelectIds = (event, value) => {
+  const [openSearch, setOpenSearch] = React.useState(false);
+
+  const handleSelectIds = (option) => {
     // setSearchBarTokenId state
-    if (value == null) {
-      value = '';
+    if (option.value == null) {
+      option.value = props.searchBarTokenId;
     }
-    props.setSearchBarTokenId(value);
+    props.setSearchBarTokenId(option.value);
     // replace the tokenId in the searchBar
     window.history.pushState(
       '',
       '',
-      window.location.href.replace(`/${props.searchBarTokenId}/`, `/${value}/`),
+      window.location.href.replace(`${props.searchBarTokenId}`, `${option.value}`),
     );
+    useRoute();
   };
 
-  const getNFTCollection = async () => {
-    console.log('tokenIDfromContext',tokenId);
+  const setTokenAndCanisterId = async () => {
+    const URL = window.location.href;
+    if (URL.includes('collection')) {
+      useRoute().then(({ canisterId }) => {
+        setCanisterId(canisterId);
+        console.log('canisterId', canisterId);
+      });
+    } else {
+      useRoute().then(({ tokenId, canisterId }) => {
+        setTokenId(tokenId);
+        setCanisterId(canisterId);
+        console.log('tokenId', tokenId);
+        console.log('canisterId', canisterId);
+      });
+    }
+  };
+
+  useEffect(() => {
+    setTokenAndCanisterId();
+  }, []);
+
+  const NFTCollection = async () => {
     setSelectTokenIds(['Loading...']);
-    const response = await actor?.collection_nft_origyn([]);
+    OrigynClient.getInstance().init(true, canisterId, { actor });
+    const response = await getNftCollectionMeta();
     const collectionNFT = response.ok;
     const obj_token_ids = collectionNFT.token_ids;
-    const number_ids = collectionNFT.token_ids_count[0].toString();
-    setIdsNumber(number_ids);
-    let x: string;
+
+    // TODO: uncomment when idsNumber is being used
+    // const number_ids = collectionNFT.token_ids_count[0].toString();
+    // setIdsNumber(number_ids);
     const arrayTokenIds = [];
-    for (x in obj_token_ids) {
+    for (var x in obj_token_ids) {
       var newID = obj_token_ids[x];
       // This is the array created to be filtered with Intersection
       arrayTokenIds.push(newID);
       setSelectTokenIds([...newID]);
     }
-
     // Check if the token Id is in the url
     const splitted_url: string[] = window.location.href.split('/');
     // Empty indexID
@@ -64,82 +96,90 @@ export const SearchbarNft = (props : any) => {
     const intersection = splitted_url.filter((e) => arrayTokenIds[0].indexOf(e) > -1);
     if (intersection.length > 0) {
       // setSearchBarTokenId state
-      props.setSearchBarTokenId(tokenId);
+      useRoute().then(({ tokenId }) => {
+        props.setSearchBarTokenId(tokenId);
+      });
+      console.log('here');
     } else {
-      // setSearchBarTokenId state
-      props.setSearchBarTokenId('Not selected');
+      if (window.location.href.includes('collection')) {
+        props.setSearchBarTokenId(obj_token_ids[0][0]);
+      } else {
+        // setSearchBarTokenId state
+        console.log('here2');
+        props.setInvalidToken(true);
+      }
     }
   };
   // if the actor changes getNftCollection is called
   useEffect(() => {
     if (actor) {
-      getNFTCollection();
+      NFTCollection();
     }
-  }, [actor]);
-
+  }, []);
   return (
-    <Box
-      component={Paper}
-      elevation={2}
-      sx={{ margin: 2, width: '100%', padding: 2 }}
-    >
-      {props.isLoading ? (
-        <Box sx={{ textAlign: 'center' }}>
-          <CircularProgress color="inherit" />
-        </Box>
-      ) : (
-        <FormControl sx={{ m: 1, width: '100%' }}>
-          {tokenId == "" ? (
-            <div>
-              <Typography
-                sx={{
-                  m: 1,
-                  width: '95%',
-                }}
-              >
-                Collection name: <b>{collectionName(tokenId)}</b>
-              </Typography>
-              <Typography
-                sx={{
-                  m: 1,
-                  borderBottom: '1px solid',
-                  paddingBottom: 2,
-                  width: '95%',
-                }}
-              >
-                Current Token ID: <b>{props.searchBarTokenId}</b>
-              </Typography>
-            </div>
+    <>
+      <>
+        <Container padding="16px">
+          {props.isLoading ? (
+            <Card type="filled" align="center" padding="16px">
+              Loading...
+            </Card>
           ) : (
-            <Typography
-              sx={{
-                m: 1,
-                borderBottom: '1px solid',
-                paddingBottom: 2,
-                width: '95%',
-              }}
-            >
-              Current Token ID: <b>{props.searchBarTokenId}</b>
-            </Typography>
+            <Container>
+              <>
+                <Grid columns={3}>
+                  <Grid column={1}>
+                    {tokenId == '' ? (
+                      <Container padding="16px">
+                        Collection name: <b>{collectionName()}</b>
+                      </Container>
+                    ) : (
+                      <></>
+                    )}
+                    <Container padding="16px">
+                      Current Token ID: <b>{props.searchBarTokenId}</b>
+                    </Container>
+                  </Grid>
+                </Grid>
+                <br />
+                <HR />
+              </>
+              <Container padding="16px">Search for NFTs</Container>
+              <Container padding="16px">
+                <Grid columns={4}>
+                  <Grid column={1}>
+                    <Flex flexFlow="row" align="center" gap={8}>
+                      <Button iconButton onClick={() => setOpenSearch(!openSearch)}>
+                        <Icons.SearchIcon width={18} height={18} />
+                      </Button>
+                      {openSearch && (
+                        <TextInput
+                          onChange={(text) => {
+                            handleSelectIds(text.target.value);
+                          }}
+                        />
+                      )}
+                      <Select
+                        placeholder="Token Ids"
+                        selectedOption={props.searchBarTokenId}
+                        handleChange={(opt) => {
+                          handleSelectIds(opt);
+                        }}
+                        options={selectTokenIds.map((token) => {
+                          return {
+                            value: token,
+                            label: token,
+                          };
+                        })}
+                      />
+                    </Flex>
+                  </Grid>
+                </Grid>
+              </Container>
+            </Container>
           )}
-          <Typography sx={{ m: 1, fontSize: 13 }}>
-            Search for other NFT&#39;S <em>(+{idsNumber}...)</em>
-          </Typography>
-          <Autocomplete
-            disablePortal
-            id="combo-box-demo"
-            options={selectTokenIds}
-            sx={{ width: '95%', m: 1 }}
-            renderInput={(params) => (
-              <TextField {...params} label="Other tokens IDS" />
-            )}
-            value={props.searchBarTokenId}
-            onChange={(event, newValue) => {
-              handleSelectIds(event, newValue);
-            }}
-          />
-        </FormControl>
-      )}
-    </Box>
+        </Container>
+      </>
+    </>
   );
 };
