@@ -1,82 +1,55 @@
 import React from 'react';
-import { Box, Typography } from '@mui/material';
-import Grid from '@mui/material/Grid';
 // Icons ICP & OGY
 import { ICPIcon } from '@dapp/common-assets';
 import { OGYIcon } from '@dapp/common-assets';
+import { Grid,Container, HR} from '@origyn-sa/origyn-art-ui';
+
 
 export const AuctionBid = (props: any) => {
   const { type_txn, buyer, amount, token, sale_id } = props.data;
 
   return (
-    <Box>
-      <Box
-        sx={{
-          padding: 1,
-          borderBottom: '1px solid',
-        }}
-      >
-        <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-          Transaction type:
-        </Typography>
-        <Typography variant="h5" gutterBottom data-testid="typeTrans">
-          {type_txn}
-        </Typography>
-        <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
+    <>
+    <Container>
+      <Container padding="16px">
+          Transaction type: 
+          <br/>
+          <span style={{color: 'grey'}}> {type_txn}</span>
+          <br/>
           Sale ID:
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          {sale_id}
-        </Typography>
-        <Grid container>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Buyer:
-            </Typography>
-            <Typography gutterBottom>{buyer.acc_principal_string}</Typography>
+          <br/>
+          <span style={{color: 'grey'}}> {sale_id}</span>
+        </Container>
+        <HR marginTop="8px" marginBottom="8px" />
+        <Grid columns={2}>
+          <Grid column={1}>
+              Buyer: <span style={{color: 'grey'}}>{buyer.acc_principal_string}</span>
           </Grid>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Amount:
-            </Typography>
-            <Typography gutterBottom>{amount}</Typography>
+          <Grid column={2}>
+              Amount: <span style={{color: 'grey'}}>{amount}</span>
           </Grid>
         </Grid>
-      </Box>
-      <Box sx={{ padding: 1 }}>
-        <Grid container spacing={2}>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Canister:
-            </Typography>
-            <Typography gutterBottom>{token.canister_string}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Fee:
-            </Typography>
-            <Typography gutterBottom>{token.fee}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Decimals:
-            </Typography>
-            <Typography gutterBottom>{token.decimal}</Typography>
+      </Container>
+      <Container padding="16px">
+        <Grid columns={2}>
+          <Grid column={1}>
+              Canister: <span style={{color: 'grey'}}>{token.canister_string}</span>
+              <br/>
+              Fee: <span style={{color: 'grey'}}>{token.fee}</span>
+         <br/>
+              Decimals: <span style={{color: 'grey'}}>{token.decimal}</span>
           </Grid>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Symbol:
-            </Typography>
-            <Typography>
-              {token.symbol === 'OGY' ? (
+          <Grid column={2}>
+              Symbol: {token.symbol === 'OGY' ? (
                 <OGYIcon className="token-symbol" />
               ) : (
                 <ICPIcon className="token-symbol" />
               )}
-            </Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Standard:
-            </Typography>
-            <Typography gutterBottom>{token.standard}</Typography>
+            
+              Standard: <span style={{color: 'grey'}}>{token.standard}</span>
           </Grid>
         </Grid>
-      </Box>
-    </Box>
+      </Container>
+    </>
   );
 };
