@@ -1,8 +1,7 @@
 import React from 'react';
-import { Box, Typography } from '@mui/material';
-import Grid from '@mui/material/Grid';
 // Icons ICP & OGY
 import { ICPIcon, OGYIcon } from '@dapp/common-assets';
+import { Grid, Container, HR } from '@origyn-sa/origyn-art-ui';
 
 export const EscrowWithdraw = (props: any) => {
   const { type_txn, buyer, seller, token, token_id, amount, trx_id } = props.data;
@@ -23,93 +22,59 @@ export const EscrowWithdraw = (props: any) => {
     id_trans = _nat.toString();
   }
   return (
-    <Box>
-      <Box
-        sx={{
-          padding: 1,
-          borderBottom: '1px solid',
-        }}
-      >
-        <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-          Transaction type:
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          {type_txn}
-        </Typography>
-        <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-          Transaction ID:
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          {id_trans}
-        </Typography>
-      </Box>
-      <Box
-        sx={{
-          padding: 1,
-          borderBottom: '1px solid',
-        }}
-      >
-        <Grid container>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Buyer:
-            </Typography>
-            <Typography gutterBottom>{buyer.acc_principal_string}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Seller:
-            </Typography>
-            <Typography gutterBottom>{seller.acc_principal_string}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Token ID:
-            </Typography>
-            <Typography gutterBottom>{token_id}</Typography>
+    <Container>
+      <Container padding="16px">
+        Transaction type: <span style={{color: 'grey'}}>{type_txn}</span>
+        <br />
+        Transaction ID:  <span style={{color: 'grey'}}>{id_trans}</span>
+      </Container>
+      <HR marginTop="8px" marginBottom="8px" />
+      <Container padding="16px">
+        <Grid columns={2}>
+          <Grid column={1}>
+            Buyer:
+            <br />
+            <span style={{color: 'grey'}}>{buyer.acc_principal_string}</span>
+            <br />
+            Seller:
+            <br />
+            <span style={{color: 'grey'}}>{seller.acc_principal_string}</span>
+            <br />
+            Token ID:  <span style={{color: 'grey'}}>{token_id}</span>
           </Grid>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Amount:
-            </Typography>
-            <Typography gutterBottom>{amount}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Fee:
-            </Typography>
-            <Typography gutterBottom>{token_fee}</Typography>
+          <Grid column={2}>
+            Amount:
+            <br />
+            <span style={{color: 'grey'}}>{amount}</span>
+            <br />
+            Fee:
+            <br />
+            <span style={{color: 'grey'}}>{token_fee}</span>
           </Grid>
         </Grid>
-      </Box>
-      <Box sx={{ padding: 1 }}>
-        <Grid container spacing={2}>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Canister:
-            </Typography>
-            <Typography gutterBottom>{token.canister_string}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Fee:
-            </Typography>
-            <Typography gutterBottom>{token.fee}</Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Decimals:
-            </Typography>
-            <Typography gutterBottom>{token.decimal}</Typography>
+      </Container>
+      <Container padding="16px">
+        <Grid columns={2}>
+          <Grid column={1}>
+            Canister:  <span style={{color: 'grey'}}>{token.canister_string}</span>
+            <br />
+            Fee:  <span style={{color: 'grey'}}>{token.fee}</span>
+            <br />
+            Decimals:  <span style={{color: 'grey'}}>{token.decimal}</span>
           </Grid>
-          <Grid item xs={6} md={6}>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Symbol:
-            </Typography>
-            <Typography>
-              {token.symbol === 'OGY' ? (
-                <OGYIcon className="token-symbol" />
-              ) : (
-                <ICPIcon className="token-symbol" />
-              )}
-            </Typography>
-            <Typography sx={{ fontSize: 14 }} color="text.secondary" gutterBottom>
-              Standard:
-            </Typography>
-            <Typography gutterBottom>{token.standard}</Typography>
+          <Grid column={2}>
+            Symbol:{' '}
+            {token.symbol === 'OGY' ? (
+              <OGYIcon className="token-symbol" />
+            ) : (
+              <ICPIcon className="token-symbol" />
+            )}
+            <br />
+            Standard:  <span style={{color: 'grey'}}>{token.standard}</span>
           </Grid>
         </Grid>
-      </Box>
-    </Box>
+      </Container>
+      <HR marginTop="8px" marginBottom="8px" />
+    </Container>
   );
 };
