@@ -1,6 +1,5 @@
 import React from 'react';
-import { Box, Typography } from '@mui/material';
-import Grid from '@mui/material/Grid';
+import { Grid, Container, HR } from '@origyn-sa/origyn-art-ui';
 // Icons ICP & OGY
 import {ICPIcon, OGYIcon} from '@dapp/common-assets';
 
@@ -14,125 +13,52 @@ export const SaleEnded = (props : any) => {
 
   return (
 
-    <Box>
-      <Box
-        sx={{
-          padding: 1,
-          borderBottom: '1px solid',
-        }}
-      >
-        <Typography
-          sx={{ fontSize: 14 }}
-          color="text.secondary"
-          gutterBottom
-        >
-          Transaction type:
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          {props.data.type_txn}
-        </Typography>
-        <Typography
-          sx={{ fontSize: 14 }}
-          color="text.secondary"
-          gutterBottom
-        >
-          Sale ID:
-        </Typography>
-        <Typography variant="h5" gutterBottom>
-          {props.data.sale}
-        </Typography>
-      </Box>
-      <Box
-        sx={{
-          padding: 1,
-          borderBottom: '1px solid',
-        }}
-      >
-        <Typography
-          sx={{ fontSize: 14 }}
-          color="text.secondary"
-          gutterBottom
-        >
-          Amount:
-        </Typography>
-        <Typography gutterBottom>{amount}</Typography>
-        <Grid container>
-          <Grid item xs={6} md={6}>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Buyer:
-            </Typography>
-            <Typography gutterBottom>{buyer.acc_principal_string}</Typography>
+    <Container>
+      <Container padding="16px">
+        
+          Transaction type:<br/>
+          <span style={{color: 'grey'}}>{props.data.type_txn}</span>
+          <br/>
+          Sale ID:<br/>
+          <span style={{color: 'grey'}}>{props.data.sale}</span>
+        
+      </Container>
+      <HR marginTop="8px" marginBottom="8px" />
+      <Container padding="16px">
+        <Container>
+        Amount:<br/>
+        <span style={{color: 'grey'}}>{amount}</span>
+        </Container>
+        <Grid columns>
+          <Grid column={1}>
+            
+              Buyer:<br/> <span style={{color: 'grey'}}>{buyer.acc_principal_string}</span>
           </Grid>
-          <Grid item xs={6} md={6}>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Seller:
-            </Typography>
-            <Typography gutterBottom>{seller.acc_principal_string}</Typography>
+          <Grid column={2}>
+            
+              Seller:<br/> <span style={{color: 'grey'}}>{seller.acc_principal_string}</span>
           </Grid>
         </Grid>
-      </Box>
-      <Box sx={{ padding: 1 }}>
-        <Grid container spacing={2}>
-          <Grid item xs={6} md={6}>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Canister:
-            </Typography>
-            <Typography gutterBottom>{token.canister_string}</Typography>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Fee:
-            </Typography>
-            <Typography gutterBottom>{token.fee}</Typography>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Decimals:
-            </Typography>
-            <Typography gutterBottom>{token.decimal}</Typography>
+      </Container>
+      <Container>
+        <Grid columns={2}>
+          <Grid column={1}>
+            Canister:<br/> <span style={{color: 'grey'}}>{token.canister_string}</span><br/>
+             Fee:<br/> <span style={{color: 'grey'}}>{token.fee}</span><br/>
+            Decimals: <br/> <span style={{color: 'grey'}}>{token.decimal}</span><br/>
           </Grid>
-          <Grid item xs={6} md={6}>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Symbol:
-            </Typography>
-            <Typography>
-              {token.symbol === 'OGY' ? (
+          <Grid column={2}>
+             Symbol:<br/>{token.symbol === 'OGY' ? (
                 <OGYIcon className="token-symbol" />
               ) : (
                 <ICPIcon className="token-symbol" />
               )}
-            </Typography>
-            <Typography
-              sx={{ fontSize: 14 }}
-              color="text.secondary"
-              gutterBottom
-            >
-              Standard:
-            </Typography>
-            <Typography gutterBottom>{token.standard}</Typography>
+            <br/>
+            
+              Standard:<br/> <span style={{color: 'grey'}}>{token.standard}</span>
           </Grid>
         </Grid>
-      </Box>
-    </Box>
+      </Container>
+    </Container>
   );
 };
