@@ -1,42 +1,36 @@
-import React, { Fragment, useState } from 'react';
-import {
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  Box,
-  Tabs,
-  Tab,
-  TextField,
-  Button,
-  List,
-  ListItem,
-  Checkbox,
-  ListItemButton,
-  ListItemText,
-} from '@mui/material';
+import React, { useContext, useState } from 'react';
 import { useSnackbar } from 'notistack';
 import { useTokensContext } from '@dapp/features-tokens-provider';
 import { IdlStandard, isLocal } from '@dapp/utils';
-import { TabPanel } from '../TabPanel';
 import { TokenIcon } from '../TokenIcon';
 import { LoadingContainer } from '../LoadingContainer';
-import CircularProgress from '@mui/material/CircularProgress';
-import { useAuthContext } from '@dapp/features-authentication';
+import {
+  Button,
+  Card,
+  CheckboxInput,
+  Container,
+  Flex,
+  Modal,
+  Select,
+  TabContent,
+  TextInput,
+} from '@origyn-sa/origyn-art-ui';
+import { AuthContext } from '../../../authentication';
 
 export const WalletTokens = ({ children }: any) => {
-  const { principal, localDevelopment } = useAuthContext();
+  const { principal } = useContext(AuthContext);
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-  const [selectedTab, setSelectedTab] = useState<number>(0);
+  //const [selectedTab, setSelectedTab] = useState<number>(0)
   const [selectedStandard, setSelectedStandard] = useState<string>(IdlStandard.DIP20.toString());
   const [inputCanisterId, setInputCanisterId] = useState<string>('');
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const { tokens, addToken, toggleToken, refreshAllBalances } = useTokensContext();
+  const { tokens, addToken, toggleToken } = useTokensContext();
   const { enqueueSnackbar } = useSnackbar();
   const handleAddButton = async () => {
     if (isLoading) return;
     setIsLoading(true);
     const tokenResponse = await addToken(
-      isLocal() && localDevelopment,
+      isLocal(),
       inputCanisterId,
       IdlStandard[selectedStandard],
       principal,
@@ -66,121 +60,97 @@ export const WalletTokens = ({ children }: any) => {
 
   const handleModalOpen = () => {
     setIsModalOpen(true);
-    refreshAllBalances(isLocal() && localDevelopment, principal);
+    // refreshAllBalances(isLocal(), principal)
   };
-  const handleTabChange = (event: React.SyntheticEvent, tab: number) => {
-    setSelectedTab(tab);
-  };
+  // const handleTabChange = (event: React.SyntheticEvent, tab: number) => {
+  //   setSelectedTab(tab)
+  // }
   const onTokenCheck = (symbol: string) => {
     toggleToken(symbol);
   };
-  const a11yProps = (index: number) => ({
-    id: `simple-tab-${index}`,
-    'aria-controls': `simple-tabpanel-${index}`,
-  });
+  // const a11yProps = (index: number) => ({
+  //   id: `simple-tab-${index}`,
+  //   'aria-controls': `simple-tabpanel-${index}`,
+  // });
   return (
     <>
-      <Dialog
-        open={isModalOpen}
-        onClose={() => handleModalClose()}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-      >
-        <DialogTitle id="alert-dialog-title">Add Tokens</DialogTitle>
-        <DialogContent>
-          <Box sx={{ borderBottom: 1, borderColor: 'divider' }}>
-            <Tabs value={selectedTab} onChange={handleTabChange} aria-label="basic tabs example">
-              <Tab label="Current Tokens" {...a11yProps(0)} />
-              <Tab label="Custom Token" {...a11yProps(1)} />
-            </Tabs>
-          </Box>
-          <TabPanel value={selectedTab} index={0} padding={1}>
-            <List dense disablePadding>
-              {Object.keys(tokens).map((key: string) => {
-                const token = tokens[key];
-                const labelId = `checkbox-list-secondary-label-${token.symbol}`;
-                return (
-                  <ListItem
-                    key={`${token.symbol}-${token.enabled}`}
-                    secondaryAction={
-                      <Checkbox
-                        edge="end"
-                        onChange={() => onTokenCheck(token.symbol)}
-                        checked={token.enabled}
-                        inputProps={{ 'aria-labelledby': labelId }}
-                      />
-                    }
-                    disablePadding
-                  >
-                    <ListItemButton>
-                      <TokenIcon symbol={token.icon} />
-                      <ListItemText
-                        id={labelId}
-                        style={{ marginLeft: '5px' }}
-                        primary={token.symbol}
-                      />
-                      {token.balance > -1 ? token.balance : <CircularProgress size={10} />}
-                    </ListItemButton>
-                  </ListItem>
-                );
-              })}
-            </List>
-          </TabPanel>
-          <TabPanel value={selectedTab} index={1}>
-            <div
-              style={{
-                display: 'flex',
-                flexWrap: 'wrap',
-                justifyContent: 'center',
-                gap: '1.5rem',
-                opacity: isLoading ? '0.4' : '1',
-              }}
-            >
-              <TextField
-                style={{ width: '100%' }}
-                id="standard-helperText"
-                label="Canister Id"
-                helperText="The canister id of the ledger"
-                value={inputCanisterId}
-                onChange={(e) => setInputCanisterId(e.target.value)}
-                required
-              />
-              <TextField
-                style={{ width: '100%' }}
-                id="standard-select-currency-native"
-                select
-                label="Token Standard"
-                value={selectedStandard}
-                onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
-                  setSelectedStandard(event.target.value);
-                }}
-                SelectProps={{
-                  native: true,
-                }}
-                helperText="Select the standard of the token"
-              >
-                {Object.keys(IdlStandard)
-                  .filter((standard) => isNaN(parseInt(standard)))
-                  .map((standard) => (
-                    <option key={standard} value={standard}>
-                      {standard}
-                    </option>
-                  ))}
-              </TextField>
-              <br />
-              <Button variant="contained" onClick={handleAddButton}>
-                Add token
-              </Button>
-            </div>
-          </TabPanel>
+      <Modal isOpened={isModalOpen} closeModal={() => handleModalClose()} size="md">
+        <Container size="full" padding="48px">
+          <h2>Manage Tokens</h2>
+          <br />
+          <TabContent
+            fullWidth
+            borderBottom
+            tabs={[
+              { title: 'Current Tokens', id: 'Current Tokens' },
+              { title: 'Custom Token', id: 'Custom Token' },
+            ]}
+            content={[
+              <Flex flexFlow="column" gap={18} fullWidth key="tabContentCol1">
+                <br />
+                {Object.keys(tokens).map((key: string) => {
+                  const token = tokens[key];
+                  // const labelId = `checkbox-list-secondary-label-${token.symbol}`;
+                  return (
+                    <Card
+                      key={`${token.symbol}-${token.enabled}`}
+                      justify="space-between"
+                      align="center"
+                      padding="16px"
+                    >
+                      <Flex gap={8} align="center">
+                        <CheckboxInput
+                          name={token.symbol}
+                          onChange={() => onTokenCheck(token.symbol)}
+                          checked={token.enabled}
+                        />
+                        <TokenIcon symbol={token.icon} />
+                        <b>{token.symbol}</b>
+                      </Flex>
+                      <div>{token.balance}</div>
+                    </Card>
+                  );
+                })}
+              </Flex>,
+              <Flex flexFlow="column" gap={18} fullWidth key="tabContentCol2">
+                <br />
+                <TextInput
+                  id="standard-helperText"
+                  label="Canister Id"
+                  value={inputCanisterId}
+                  onChange={(e) => setInputCanisterId(e.target.value)}
+                  required
+                />
+                <Select
+                  label="Token Standard"
+                  handleChange={(option) => {
+                    setSelectedStandard(option.value);
+                  }}
+                  options={Object.keys(IdlStandard)
+                    .filter((standard) => isNaN(parseInt(standard)))
+                    .map((standard) => ({
+                      value: standard,
+                      label: standard,
+                    }))}
+                />
+                <Flex justify="flex-end">
+                  <Button btnType="outlined" onClick={handleAddButton}>
+                    Add token
+                  </Button>
+                </Flex>
+              </Flex>,
+            ]}
+          />
           {isLoading && (
             <div style={{ marginTop: 5 }}>
               <LoadingContainer />
             </div>
           )}
-        </DialogContent>
-      </Dialog>
-      <div onClick={handleModalOpen}>{children}</div>
+        </Container>
+      </Modal>
+      <Button btnType="filled" onClick={handleModalOpen}>
+        {children}
+      </Button>
     </>
   );
 };
