@@ -1,17 +1,15 @@
-export * from './Layout'
-export * from './LoadingContainer'
-export * from './Logo'
-export * from './NatPrice'
-export * from './Preloader'
-export * from './Table'
-export * from './TabPanel'
-export * from './TokenIcon'
-export * from './WalletTokens'
-export * from './SearchbarNft'
-export * from './TransactionFilter'
-export * from './TransactionModal'
-export * from './TransactionsTable'
-export * from './VLabel'
-
-
-
+export * from './Layout';
+export * from './LoadingContainer';
+export * from './Logo';
+export * from './NatPrice';
+export * from './Preloader';
+export * from './Table';
+export * from './TabPanel';
+export * from './TokenIcon';
+export * from './WalletTokens';
+export * from './SearchbarNft';
+export * from './TransactionFilter';
+export * from './TransactionModal';
+export * from './TransactionsTable';
+export * from './VLabel';
+export * from './SwitchCanisterCollection';
