@@ -1,8 +1,6 @@
-export const collectionName = (_tokenId: string) => {
-  if (_tokenId == '') {
+export const collectionName = () => {
     //find the id of 'collection' -1
     const url_array: string[] = window.location.pathname.split('/');
     const collection_id: number = url_array.indexOf('collection') - 1;
     return url_array[collection_id];
-  }
-};
+  };
